requirement.txt
alabaster==0.7.12
attrs==19.1.0
Automat==0.7.0
Babel==2.7.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
docutils==0.15.2
fastecdsa==1.7.5
hyperlink==19.0.0
idna==2.8
imagesize==1.1.0
incremental==17.5.0
Jinja2==2.10.3
MarkupSafe==1.1.1
packaging==19.2
pycparser==2.19
Pygments==2.4.2
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
pyparsing==2.4.2
pytz==2019.3
requests==2.22.0
six==1.12.0
snowballstemmer==2.0.0
Sphinx==2.2.0
sphinx-minoo-theme==0.9.7
sphinx-theme-pd==0.8.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
termcolor==1.1.0
Twisted==19.7.0
urllib3==1.25.6
zope.interface==4.6.0